pyproject.toml
[project]
name = "matching"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic[email]>=2.11.4",
    "pydantic-ai>=0.1.9",
    "pymupdf>=1.25.5",
    "pillow>=11.2.1",
    "spacy>=3.8.5",
    "skillner>=1.0.3",
    "ipython>=8.18.1",
    "psutil>=7.0.0",
    "sentence-transformers>=4.1.0",
    "numpy>=2.0.2",
]
